.reviewer/pyproject.toml
[project]
name = "reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
